requirements.txt
django
django-tastypie
pandas
sklearn
gunicorn
django-heroku